sandbox - move diagnostics service to electron sandbox
