+++ src/vs/workbench/services/diagnostics/electron-sandbox/diagnosticsService.ts
@@ -6,7 +6,7 @@
 import { createChannelSender } from 'vs/base/parts/ipc/common/ipc';
 import { ISharedProcessService } from 'vs/platform/ipc/electron-sandbox/sharedProcessService';
 import { registerSingleton } from 'vs/platform/instantiation/common/extensions';
-import { IDiagnosticsService } from 'vs/platform/diagnostics/node/diagnosticsService';
+import { IDiagnosticsService } from 'vs/platform/diagnostics/common/diagnostics';
 
 // @ts-ignore: interface is implemented via proxy
 export class DiagnosticsService implements IDiagnosticsService {
